make ta tk more intuitive

+++ .zshrc
@@ -185,7 +185,8 @@ tn() {
 # ta [query...]: fuzzy-pick a session to attach to (auto-selects if query matches exactly one)
 ta() {
     local session
-    session=$(tmux list-sessions -F '#{session_name}#{?session_attached, (attached),}' | fzf -q "$*" --select-1 --exit-0 | sed 's/ (attached)$//') || return
+    session=$(tmux list-sessions -F '#{session_name}#{?session_attached, (attached),}' | fzf -q "$*" --select-1 --exit-0 --reverse | sed 's/ (attached)$//')
+    [[ -z "$session" ]] && return
     if [[ -n "$TMUX" ]]; then
         tmux switch-client -t "$session"
     else
@@ -195,7 +196,8 @@ ta() {
 # tk [query...]: fuzzy-pick a session to kill (always shows picker, even with exact match)
 tk() {
     local session
-    session=$(tmux list-sessions -F '#{session_name}#{?session_attached, (attached),}' | fzf -q "$*" --exit-0 | sed 's/ (attached)$//') || return
+    session=$(tmux list-sessions -F '#{session_name}#{?session_attached, (attached),}' | fzf -q "$*" --exit-0 --reverse | sed 's/ (attached)$//')
+    [[ -z "$session" ]] && return
     tmux kill-session -t "$session"
 }
 # runall <cmd...>: send <cmd> + Enter to every zsh pane across all tmux sessions
